+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 -   Update membership addition label when changing language [BKDO-3337](https://bkdev.atlassian.net/browse/BKDO-3337)
 -   Update function labels [BKDO-3393](https://bkdev.atlassian.net/browse/BKDO-3393)
 -   Mark people with expired memberships [BKDO-3399](https://bkdev.atlassian.net/browse/BKDO-3399)
+-   Display warning when requirements profile is required [BKDO-3399](https://bkdev.atlassian.net/browse/BKDO-3399)
 
 ## 1.2.0